Create delete and modify work perfectly

+++ ProjetWeb/view/myAd.php
@@ -28,14 +28,21 @@ ob_start();
                         <?php foreach($articles as $article) :?>
                             <div class="s-12 m-12 l-4 margin-m-bottom div-row">
                                 <img src="<?=$article['inputPictures'];?>">
-                                <a href="../index.php?action=adDetails&ID=<?= $article['ID']; ?>"><br><strong><?=$article['inputNameAnnonce']; ?></strong><br></a>                                <a><?=$article['inputPrice']; ?> CHF</a><br>
+                                <a href="../index.php?action=adDetails&ID=<?= $article['ID']; ?>"><br><strong><?=$article['inputNameAnnonce']; ?></strong><br></a>
+                                <a><?=$article['inputPrice']; ?> CHF</a><br>
                                 <a><?=$article['inputCity']; ?></a><br><br>
+
+
+
                                 <a class="text-more-info text-primary-hover" href="../index.php?action=modifAd&ID=<?= $article['ID']; ?>">Modifier</a>
-                                <a class="text-more-info text-primary-hover" href="../index.php?action=deleteArticle&ID=<?= $article['ID']; ?>">Supprimer</a>
+                                <?php if($article['active'] == true): ?>
+                                    <a class="text-more-info text-primary-hover" href="../index.php?action=deleteArticle&ID=<?= $article['ID']; ?>">Supprimer</a>
+                                <?php else: ?>
+                                    <a> Deleted</a>
+                                <?php endif; ?>
+
                             </div>
                         <?php endforeach; ?>
-
-
                     </div>
 
                 </div>
